requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
numpy
python-dotenv
huggingface_hub
accelerate
PyPDF2
python-docx
openpyxl
pytesseract
Pillow
fpdf
matplotlib
pandas
streamlit
plotly
SpeechRecognition
